refactor all sub components; implement proficiencies
