+++ client/src/components/CharacterPage/Proficiencies/style.css
@@ -1,0 +1,9 @@
+.section-title.proficiencies {
+  margin-left: 240px
+}
+
+@media screen and (max-width: 992px) {
+  .section-title.proficiencies {
+    margin-left: 0;
+  }
+}
