+++ client/src/components/CharacterPage/SpellSlots/index.jsx
@@ -1,18 +1,18 @@
 import "./style.css";
 import PropTypes from "prop-types";
-import { Character } from "../../models/Character";
+import { Character } from "../../../models/Character";
 
-export default function Equipment(props) {
+export default function SpellSlots(props) {
   const character = new Character(props.character);
 
   return (
     <div className="fs-3">
       <div className="character-view-header sticky-top pt-1">
-        <div className="d-flex" role="button" onClick={() => props.toggleSectionShowing()} data-bs-toggle="collapse" data-bs-target="#character-view-equipment" aria-expanded="false" aria-controls="character-view-equipment">
+        <div className="d-flex" role="button" onClick={() => props.toggleSectionShowing()} data-bs-toggle="collapse" data-bs-target="#character-view-spell-slots" aria-expanded="false" aria-controls="character-view-spell-slots">
           <h2 className="section-title">
-            Equipment
+            Spell Slots
           </h2>
-          {props.isShowingEquipment ?
+          {props.isShowingSpellSlots ?
             <i className="bi bi-chevron-down fs-3 px-3" aria-label="chevron-down"></i> :
             <i className="bi bi-chevron-up fs-3 px-3" aria-label="chevron-up"></i>
           }
@@ -21,15 +21,15 @@ export default function Equipment(props) {
         <button className="btn btn-secondary button-edit">Edit</button>
       </div>
 
-      <div id="character-view-equipment" className="collapse show">
+      <div id="character-view-spell-slots" className="collapse show">
         PLACEHOLDER
       </div>
     </div>
   )
 }
 
-Equipment.propTypes = {
+SpellSlots.propTypes = {
   character: PropTypes.object,
   toggleSectionShowing: PropTypes.func,
-  isShowingEquipment: PropTypes.bool,
+  isShowingSpellSlots: PropTypes.bool,
 }
