+++ client/src/components/CharacterPage/Spells/index.jsx
@@ -1,18 +1,18 @@
 import "./style.css";
 import PropTypes from "prop-types";
-import { Character } from "../../models/Character";
+import { Character } from "../../../models/Character";
 
-export default function SpellSlots(props) {
+export default function Spells(props) {
   const character = new Character(props.character);
 
   return (
     <div className="fs-3">
       <div className="character-view-header sticky-top pt-1">
-        <div className="d-flex" role="button" onClick={() => props.toggleSectionShowing()} data-bs-toggle="collapse" data-bs-target="#character-view-spell-slots" aria-expanded="false" aria-controls="character-view-spell-slots">
+        <div className="d-flex" role="button" onClick={() => props.toggleSectionShowing()} data-bs-toggle="collapse" data-bs-target="#character-view-spells" aria-expanded="false" aria-controls="character-view-spells">
           <h2 className="section-title">
-            Spell Slots
+            Spells
           </h2>
-          {props.isShowingSpellSlots ?
+          {props.isShowingSpells ?
             <i className="bi bi-chevron-down fs-3 px-3" aria-label="chevron-down"></i> :
             <i className="bi bi-chevron-up fs-3 px-3" aria-label="chevron-up"></i>
           }
@@ -20,16 +20,16 @@ export default function SpellSlots(props) {
 
         <button className="btn btn-secondary button-edit">Edit</button>
       </div>
-
-      <div id="character-view-spell-slots" className="collapse show">
+      
+      <div id="character-view-spells" className="collapse show">
         PLACEHOLDER
       </div>
     </div>
   )
 }
 
-SpellSlots.propTypes = {
+Spells.propTypes = {
   character: PropTypes.object,
   toggleSectionShowing: PropTypes.func,
-  isShowingSpellSlots: PropTypes.bool,
+  isShowingSpells: PropTypes.bool,
 }
