+++ client/src/components/CharacterPage/Treasures/index.jsx
@@ -1,6 +1,6 @@
 import "./style.css";
 import PropTypes from "prop-types";
-import { Character } from "../../models/Character";
+import { Character } from "../../../models/Character";
 
 export default function Treasures(props) {
   const character = new Character(props.character);
