+++ client/src/components/CharacterPage/index.jsx
@@ -9,15 +9,15 @@ import Background from "./Background";
 import CharacterInfo from "./CharacterInfo";
 import AbilityScores from "./AbilityScores";
 import Weapons from "./Weapons";
-import Treasures from "./treasures";
-import SpellSlots from "./spellSlots";
-import Spells from "./spells";
+import Treasures from "./Treasures";
+import SpellSlots from "./SpellSlots";
+import Spells from "./Spells";
 import Skills from "./Skills";
 import SavingThrows from "./SavingThrows";
-import Proficiencies from "./proficiencies";
-import Languages from "./languages";
+import Proficiencies from "./Proficiencies";
+import Languages from "./Languages";
 import FeaturesTraits from "./FeaturesTraits";
-import Equipment from "./equipment";
+import Equipment from "./Equipment";
 import { useEffect, useState } from "react";
 import { SECTION_TITLE } from "../../utils/enums";
